refactor(pdf): extract private fidelity boundary tests
