+++ tools/pdf-private-fidelity.test.mjs
@@ -1,19 +1,4 @@
-import {
-  chmodSync,
-  mkdirSync,
-  mkdtempSync,
-  readFileSync,
-  readdirSync,
-  rmSync,
-  statSync,
-  symlinkSync,
-  writeFileSync,
-} from 'node:fs'
-import { homedir, tmpdir } from 'node:os'
-import { join, resolve } from 'node:path'
-import { spawnSync } from 'node:child_process'
 import { createHash } from 'node:crypto'
-import { fileURLToPath } from 'node:url'
 import { describe, expect, it } from 'vitest'
 import {
   canonicalHyphenEvidenceSha256,
@@ -28,33 +13,20 @@ import {
   comparePrivateFidelityBaseline,
   comparePrivateFidelityReceipts,
   createPrivateFidelityReceipt,
-  createPrivateFidelityRunReceipt,
   createPrivateReconstructionEvidence as createPrivateReconstructionEvidenceRaw,
   parsePrivateFidelityArguments,
-  prepareOwnerOnlyDirectory,
-  writeExclusive,
 } from './pdf-private-fidelity.mjs'
 import * as privateFidelity from './pdf-private-fidelity.mjs'
-
-const hash = 'a'.repeat(64)
-const privateCompletenessPolicy = {
-  minimumTextCoverage: 0.98,
-  minimumAssetCoverage: 1,
-  minimumRelationshipCoverage: 1,
-  maximumUnresolvedObjects: 0,
-  maximumOcrRequiredPages: 0,
-  maximumReadingOrderDiagnostics: 0,
-}
-
-function createPrivateReconstructionEvidence(
+import {
+  acceptedBaselineSha256,
+  artifact,
+  createPrivateReconstructionEvidence,
+  fidelityReceipt,
+  hash,
+  privateCompletenessPolicy,
   reconstruction,
-  artifactProjections = { 'readable-fallback': reconstruction },
-) {
-  return createPrivateReconstructionEvidenceRaw(
-    reconstruction,
-    artifactProjections,
-  )
-}
+  run,
+} from './pdf-private-fidelity.test-fixtures.mjs'
 
 function canonicalHyphenDeletionRecord() {
   const joinedWord = 'representation'
@@ -231,163 +203,12 @@ function crossReferenceRelationship(overrides = {}) {
   }
 }
 
-function reconstruction(ready = true, ledgerAvailable = true) {
-  return {
-    source: { pageCount: 3 },
-    completeness: {
-      sourceTextCharacters: 100,
-      outputTextCharacters: ready ? 100 : 90,
-      matchedTextCharacters: ready ? 100 : 90,
-      textCoverage: ready ? 1 : 0.9,
-      duplicateCanonicalSpanCount: 0,
-      missingSourceRegionCount: 0,
-      unprovenancedRenderedUnitCount: 0,
-      expectedInlineSpanCount: 1,
-      mappedInlineSpanCount: 1,
-      inlineSpanCoverage: 1,
-      expectedHyperlinkCount: 1,
-      mappedHyperlinkCount: 1,
-      hyperlinkCoverage: 1,
-      lineBoundaryCount: ledgerAvailable ? 1 : 0,
-      decidedLineBoundaryCount: ledgerAvailable ? 1 : 0,
-      unresolvedCorruptingJoinCount: 0,
-      structurallyConsumedLineBoundaryCount: 0,
-      sourceAssetCount: 1,
-      exportedAssetCount: 1,
-      assetCoverage: 1,
-      expectedRelationshipCount: 1,
-      resolvedRelationshipCount: 1,
-      relationshipCoverage: 1,
-      unresolvedObjectCount: ready ? 0 : 1,
-      unresolvedObjects: {
-        assets: 0,
-        captions: ready ? 0 : 1,
-        tables: 0,
-        equations: 0,
-        citations: 0,
-        footnoteReferences: 0,
-        footnotes: 0,
-      },
-      ocrRequiredPages: [],
-      readingOrderDiagnostics: 0,
-      readingOrderEvaluation: {
-        schemaVersion: '1.0.0',
-        algorithm: 'deterministic-geometry-v1',
-        mode: 'deterministic-only',
-        regionCount: 2,
-        acceptedEdgeCount: 1,
-        unresolvedEdgeCount: 0,
-        cycleRate: 0,
-        orderAccuracy: null,
-        provider: null,
-        modelVersion: null,
-        latencyMs: 0,
-        costUsd: 0,
-        reviewRequired: false,
-      },
-    },
-    readiness: {
-      ready,
-      status: ready ? 'ready' : 'review-required',
-      policy: privateCompletenessPolicy,
-      blockingDiagnosticCodes: ready ? [] : ['INCOMPLETE_TEXT_COVERAGE'],
-    },
-    paper: {
-      id: 'paper-v1',
-      version: '1.0.0',
-      status: 'working',
-      title: 'Synthetic paper',
-      subtitle: 'Private runner fixture',
-      authors: ['Example Author'],
-      updated: '2026-07-20',
-      abstract: 'Synthetic abstract.',
-      nodes: [
-        {
-          id: 'heading-1',
-          source: 'region-1',
-          type: 'heading',
-          level: 1,
-          text: 'Synthetic paper',
-        },
-        {
-          id: 'paragraph-1',
-          source: 'region-2',
-          type: 'paragraph',
-          text: 'Synthetic body.',
-        },
-      ],
-    },
-    readingOrder: { order: ['region-1', 'region-2'] },
-    regions: [
-      {
-        id: 'region-1',
-        page: 1,
-        lines: [{ id: 'line-1' }, { id: 'line-2' }],
-      },
-      { id: 'region-2', page: 1, lines: [{ id: 'line-3' }] },
-    ],
-    canonicalHyphenBoundaryDecisions: [],
-    canonicalHyphenBoundaryDecisionCount: 0,
-    visualRelationships: [{ kind: 'figure', status: 'matched' }],
-    noteRelationships: [],
-    assets: [],
-    diagnostics: ready
-      ? [{ severity: 'info', code: 'SOURCE_OK' }]
-      : [
-          {
-            severity: 'error',
-            code: 'INCOMPLETE_TEXT_COVERAGE',
-          },
-        ],
-    ...(ledgerAvailable
-      ? {
-          lineBoundaryDecisions: [
-            {
-              id: 'transition-1',
-              page: 1,
-              regionId: 'region-1',
-              fromLineId: 'line-1',
-              toLineId: 'line-2',
-              outcome: 'space',
-              evidence: [],
-            },
-          ],
-          unresolvedCorruptingJoinCount: 0,
-          structurallyConsumedLineBoundaryCount: 0,
-        }
-      : {}),
-  }
-}
-
 function requireReconstructionReview(source, code) {
   source.readiness.ready = false
   source.readiness.status = 'review-required'
   source.readiness.blockingDiagnosticCodes = [code]
   source.diagnostics.push({ severity: 'error', code })
   return source
-}
-
-function artifact(target, suffix = '', source = reconstruction()) {
-  const parity =
-    createPrivateReconstructionEvidence(source).artifactParity.publication
-  const evidence = {
-    target,
-    profileVersion: '1.0.0',
-    mode: 'publication',
-    byteLength: 1024,
-    sha256: `${'b'.repeat(63)}${suffix || 'b'}`,
-    canonicalNodeCount: parity.canonicalNodeCount,
-    canonicalNodeSequenceSha256: parity.canonicalNodeSequenceSha256,
-    canonicalContentSha256: parity.canonicalContentSha256,
-    relationshipCount: parity.relationshipCount,
-    relationshipGraphSha256: parity.relationshipGraphSha256,
-    assetCount: parity.assetCount,
-    assetManifestSha256: parity.assetManifestSha256,
-    inlineSemanticLedger: parity.inlineSemanticLedger,
-    structuralValidation: 'passed',
-    epubCheck: { status: 'skipped', reason: 'not-required' },
-  }
-  return { ...evidence, receiptSha256: canonicalJsonHash(evidence) }
 }
 
 function inspectedArtifact(source, content) {
@@ -413,59 +234,6 @@ function inspectedArtifact(source, content) {
       },
     },
   )
-}
-
-function run(ordinal, source, artifacts, readableFallbackProjection = source) {
-  return createPrivateFidelityRunReceipt({
-    ordinal,
-    reconstruction: source,
-    artifacts,
-    artifactProjections: {
-      'readable-fallback': readableFallbackProjection,
-    },
-  })
-}
-
-function fidelityReceipt({
-  paperId = 'paper-v1',
-  sourceSha256 = hash,
-  byteLength = 123,
-  decisionSetSha256 = null,
-  profiles = ['mobile'],
-  repeat = 2,
-  ready = true,
-  transformReconstruction = (value) => value,
-  transformArtifact = (value) => value,
-} = {}) {
-  const runs = Array.from({ length: repeat }, (_, index) => {
-    const ordinal = index + 1
-    const source = transformReconstruction(reconstruction(ready), ordinal)
-    const artifacts = profiles.map((profile) => {
-      const candidate = transformArtifact(
-        artifact(profile, '', source),
-        profile,
-        ordinal,
-      )
-      const evidence = Object.fromEntries(
-        Object.entries(candidate).filter(([key]) => key !== 'receiptSha256'),
-      )
-      return { ...evidence, receiptSha256: canonicalJsonHash(evidence) }
-    })
-    return run(ordinal, source, artifacts)
-  })
-  return createPrivateFidelityReceipt({
-    paperId,
-    sourceSha256,
-    byteLength,
-    decisionSetSha256,
-    runs,
-    repeat,
-    profiles,
-  })
-}
-
-function acceptedBaselineSha256(receipt) {
-  return canonicalJsonHash(receipt)
 }
 
 function recomputePrivateReconstructionReceiptSha256(runReceipt) {
@@ -2650,613 +2418,4 @@ describe('private PDF fidelity runner', () => {
       ),
     ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
   })
-
-  it('rejects legacy private receipts that predate complete semantic evidence', () => {
-    const legacy = fidelityReceipt()
-    legacy.schemaVersion = '1.6.0'
-
-    expect(() =>
-      comparePrivateFidelityReceipts(
-        legacy,
-        fidelityReceipt(),
-        acceptedBaselineSha256(legacy),
-      ),
-    ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
-  })
-
-  it('rejects an invalid or locally unaccepted frozen receipt generically', () => {
-    const invalidSchema = { ...fidelityReceipt(), schemaVersion: 'invalid' }
-    const corruptHash = structuredClone(fidelityReceipt())
-    corruptHash.runs[0].reconstructionReceiptSha256 = '0'.repeat(64)
-    const locallyBlocked = fidelityReceipt({ ready: false })
-    const emptyFallbackArtifact = fidelityReceipt({
-      transformArtifact(value) {
-        return {
-          ...value,
-          mode: 'readable-fallback',
-          byteLength: 0,
-          canonicalNodeCount: 0,
-        }
-      },
-    })
-    const invalidInlineCoverage = fidelityReceipt({
-      transformReconstruction(source) {
-        source.completeness.inlineSpanCoverage = 0.5
-        return source
-      },
-    })
-    const invalidCompletenessEvidence = [
-      fidelityReceipt({
-        transformReconstruction(source) {
-          delete source.completeness.expectedHyperlinkCount
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.hyperlinkCoverage = 0.5
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.mappedHyperlinkCount = 2
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.textCoverage = 0.5
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.assetCoverage = 0.5
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.relationshipCoverage = 0.5
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.matchedTextCharacters = 101
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.exportedAssetCount = 2
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.resolvedRelationshipCount = 2
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.decidedLineBoundaryCount = 2
-          return source
-        },
-      }),
-    ]
-
-    for (const baseline of [
-      invalidSchema,
-      corruptHash,
-      locallyBlocked,
-      emptyFallbackArtifact,
-      invalidInlineCoverage,
-      ...invalidCompletenessEvidence,
-    ]) {
-      expect(() =>
-        comparePrivateFidelityReceipts(
-          baseline,
-          fidelityReceipt(),
-          acceptedBaselineSha256(baseline),
-        ),
-      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
-    }
-  })
-
-  it('rejects private readiness claims that contradict policy, completeness, diagnostics, or blockers', () => {
-    const baseline = fidelityReceipt()
-    const invalidCandidates = [
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.readiness.blockingDiagnosticCodes = ['SOURCE_ERROR']
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.diagnostics.push({
-            severity: 'error',
-            code: 'SOURCE_ERROR',
-          })
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.outputTextCharacters = 97
-          source.completeness.matchedTextCharacters = 97
-          source.completeness.textCoverage = 0.97
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.unresolvedObjectCount = 1
-          source.completeness.unresolvedObjects.assets = 1
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.ocrRequiredPages = [1]
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.completeness.readingOrderDiagnostics = 1
-          source.completeness.readingOrderEvaluation.unresolvedEdgeCount = 1
-          source.completeness.readingOrderEvaluation.reviewRequired = true
-          return source
-        },
-      }),
-      fidelityReceipt({
-        transformReconstruction(source) {
-          source.readiness.ready = false
-          source.readiness.status = 'review-required'
-          return source
-        },
-      }),
-    ]
-
-    for (const candidate of invalidCandidates) {
-      expect(() =>
-        comparePrivateFidelityReceipts(
-          baseline,
-          candidate,
-          acceptedBaselineSha256(baseline),
-        ),
-      ).toThrow('INVALID_PRIVATE_FIDELITY_BASELINE')
-    }
-  })
-
-  it('loads an external baseline without exposing its path or unexpected private fields', async () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-baseline-test-'),
-    )
-    const privatePath = join(
-      temporaryDirectory,
-      'operator frozen baseline private path.json',
-    )
-    const privateText = 'PRIVATE SOURCE TEXT MUST NEVER ESCAPE'
-    try {
-      const baseline = fidelityReceipt()
-      const expectedBaselineSha256 = acceptedBaselineSha256(baseline)
-      writeFileSync(privatePath, JSON.stringify(baseline))
-      const comparison = await comparePrivateFidelityBaseline(
-        privatePath,
-        fidelityReceipt(),
-        expectedBaselineSha256,
-      )
-      const serialized = JSON.stringify(comparison)
-      expect(serialized).not.toContain(privatePath)
-      expect(serialized).not.toContain('operator frozen baseline')
-
-      writeFileSync(privatePath, JSON.stringify({ sourceText: privateText }))
-      let failure
-      try {
-        await comparePrivateFidelityBaseline(
-          privatePath,
-          fidelityReceipt(),
-          expectedBaselineSha256,
-        )
-      } catch (error) {
-        failure = String(error)
-      }
-      expect(failure).toBe('Error: INVALID_PRIVATE_FIDELITY_BASELINE')
-      expect(failure).not.toContain(privatePath)
-      expect(failure).not.toContain(privateText)
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  })
-
-  it('requires a new external owner-only directory and exclusive artifact writes', async () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-output-test-'),
-    )
-    const output = join(temporaryDirectory, 'new-proof')
-    const linkedOutput = join(temporaryDirectory, 'linked-proof')
-    try {
-      const created = await prepareOwnerOnlyDirectory(output)
-      expect(statSync(created).mode & 0o077).toBe(0)
-      await expect(prepareOwnerOnlyDirectory(created)).rejects.toThrow(
-        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
-      )
-      symlinkSync(created, linkedOutput)
-      await expect(prepareOwnerOnlyDirectory(linkedOutput)).rejects.toThrow(
-        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
-      )
-      await expect(prepareOwnerOnlyDirectory('/')).rejects.toThrow(
-        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
-      )
-      await expect(prepareOwnerOnlyDirectory(homedir())).rejects.toThrow(
-        'OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY',
-      )
-      await expect(
-        prepareOwnerOnlyDirectory(resolve('private-proof-inside-repository')),
-      ).rejects.toThrow('OUTPUT_MUST_BE_NEW_EXTERNAL_DIRECTORY')
-
-      const artifactPath = join(created, 'artifact.epub')
-      await writeExclusive(artifactPath, new Uint8Array([1, 2, 3]))
-      expect(statSync(artifactPath).mode & 0o077).toBe(0)
-      await expect(
-        writeExclusive(artifactPath, new Uint8Array([4, 5, 6])),
-      ).rejects.toMatchObject({ code: 'EEXIST' })
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  })
-
-  it('does not expose an environment-sourced path when execution fails', () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-fidelity-test-'),
-    )
-    const privatePath = join(temporaryDirectory, 'operator source.pdf')
-    const outputDirectory = join(temporaryDirectory, 'proof')
-    try {
-      const result = spawnSync(
-        process.execPath,
-        [
-          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
-          '--input-env',
-          'SRT_PRIVATE_TEST_PDF',
-          '--paper-id',
-          'paper-v1',
-          '--expected-size',
-          '123',
-          '--expected-sha256',
-          hash,
-          '--profiles',
-          'mobile',
-          '--repeat',
-          '2',
-          '--out',
-          outputDirectory,
-        ],
-        {
-          encoding: 'utf8',
-          env: { ...process.env, SRT_PRIVATE_TEST_PDF: privatePath },
-        },
-      )
-      const output = `${result.stdout}\n${result.stderr}`
-
-      expect(result.status).toBe(2)
-      expect(output).not.toContain(privatePath)
-      expect(output).not.toContain('operator source.pdf')
-      expect(output).toContain(
-        'Private PDF fidelity validation failed without exposing local paths or source content.',
-      )
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  })
-
-  it('emits only a safe error class, code, and message digest when diagnostic output is requested', () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-fidelity-diagnostic-test-'),
-    )
-    const privatePath = join(temporaryDirectory, 'missing owner source.pdf')
-    const outputDirectory = join(temporaryDirectory, 'proof')
-    try {
-      const result = spawnSync(
-        process.execPath,
-        [
-          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
-          '--input-env',
-          'SRT_PRIVATE_TEST_PDF',
-          '--paper-id',
-          'paper-v1',
-          '--expected-size',
-          '123',
-          '--expected-sha256',
-          hash,
-          '--profiles',
-          'mobile',
-          '--repeat',
-          '2',
-          '--out',
-          outputDirectory,
-          '--safe-error-diagnostic',
-        ],
-        {
-          encoding: 'utf8',
-          env: { ...process.env, SRT_PRIVATE_TEST_PDF: privatePath },
-        },
-      )
-      const output = `${result.stdout}\n${result.stderr}`
-      const diagnosticLine = result.stderr
-        .split('\n')
-        .find((line) => line.startsWith('{'))
-
-      expect(result.status).toBe(2)
-      expect(diagnosticLine).toBeDefined()
-      expect(JSON.parse(diagnosticLine)).toEqual({
-        errorClass: 'Error',
-        errorCode: 'ENOENT',
-        messageSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-        stage: 'source-stat',
-      })
-      expect(output).not.toContain(privatePath)
-      expect(output).not.toContain('missing owner source.pdf')
-      expect(output).not.toContain('no such file')
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  })
-
-  it('does not reinterpret an arbitrary identifier-shaped error message as a diagnostic code', () => {
-    expect(
-      privateFidelity.safePrivateFailureDiagnostic(
-        new Error('private-source.pdf'),
-        'test-stage',
-      ),
-    ).toEqual({
-      errorClass: 'Error',
-      errorCode: 'UNCLASSIFIED',
-      messageSha256: expect.stringMatching(/^[a-f0-9]{64}$/),
-      stage: 'test-stage',
-    })
-  })
-
-  it('treats EPUBCheck warnings as a pre-publication failure in required mode', () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-epubcheck-test-'),
-    )
-    const binaryDirectory = join(temporaryDirectory, 'bin')
-    const argumentsLog = join(temporaryDirectory, 'epubcheck-arguments.jsonl')
-    const outputDirectory = join(temporaryDirectory, 'proof')
-    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
-    const inputBytes = readFileSync(inputPath)
-    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
-    try {
-      mkdirSync(binaryDirectory)
-      const fakeEpubCheck = join(binaryDirectory, 'epubcheck')
-      writeFileSync(
-        fakeEpubCheck,
-        `#!/usr/bin/env node
-const { appendFileSync } = require('node:fs')
-if (process.argv.includes('--version')) process.exit(0)
-appendFileSync(process.env.EPUBCHECK_ARGUMENTS_LOG, JSON.stringify(process.argv.slice(2)) + '\\n')
-process.exit(1)
-`,
-      )
-      chmodSync(fakeEpubCheck, 0o755)
-
-      const result = spawnSync(
-        process.execPath,
-        [
-          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
-          '--input-env',
-          'SRT_PRIVATE_TEST_PDF',
-          '--paper-id',
-          'public-fixture',
-          '--expected-size',
-          String(inputBytes.byteLength),
-          '--expected-sha256',
-          inputSha256,
-          '--profiles',
-          'mobile',
-          '--repeat',
-          '2',
-          '--out',
-          outputDirectory,
-          '--require-epubcheck',
-          '--safe-error-diagnostic',
-        ],
-        {
-          encoding: 'utf8',
-          env: {
-            ...process.env,
-            EPUBCHECK_ARGUMENTS_LOG: argumentsLog,
-            PATH: `${binaryDirectory}:${process.env.PATH ?? ''}`,
-            SRT_PRIVATE_TEST_PDF: inputPath,
-          },
-          timeout: 120_000,
-        },
-      )
-      const diagnostic = JSON.parse(
-        result.stderr.split('\n').find((line) => line.startsWith('{')),
-      )
-
-      expect(result.status, result.stderr).toBe(2)
-      expect(diagnostic).toMatchObject({
-        errorClass: 'Error',
-        errorCode: 'EPUBCHECK_FAILED',
-        stage: 'epubcheck-validate',
-      })
-      expect(
-        readFileSync(argumentsLog, 'utf8')
-          .trim()
-          .split('\n')
-          .map((line) => JSON.parse(line)),
-      ).toEqual([
-        ['--failonwarnings', expect.stringMatching(/publication\.epub$/)],
-      ])
-      expect(readdirSync(outputDirectory)).toEqual([])
-      expect(`${result.stdout}\n${result.stderr}`).not.toContain(inputPath)
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  }, 120_000)
-
-  it('records every required EPUBCheck pass in the sanitized private receipt', () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-epubcheck-pass-test-'),
-    )
-    const binaryDirectory = join(temporaryDirectory, 'bin')
-    const argumentsLog = join(temporaryDirectory, 'epubcheck-arguments.jsonl')
-    const outputDirectory = join(temporaryDirectory, 'proof')
-    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
-    const inputBytes = readFileSync(inputPath)
-    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
-    try {
-      mkdirSync(binaryDirectory)
-      const fakeEpubCheck = join(binaryDirectory, 'epubcheck')
-      writeFileSync(
-        fakeEpubCheck,
-        `#!/usr/bin/env node
-const { appendFileSync } = require('node:fs')
-if (process.argv.includes('--version')) process.exit(0)
-appendFileSync(process.env.EPUBCHECK_ARGUMENTS_LOG, JSON.stringify(process.argv.slice(2)) + '\\n')
-`,
-      )
-      chmodSync(fakeEpubCheck, 0o755)
-
-      const result = spawnSync(
-        process.execPath,
-        [
-          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
-          '--input-env',
-          'SRT_PRIVATE_TEST_PDF',
-          '--paper-id',
-          'public-fixture',
-          '--expected-size',
-          String(inputBytes.byteLength),
-          '--expected-sha256',
-          inputSha256,
-          '--profiles',
-          'mobile',
-          '--repeat',
-          '2',
-          '--out',
-          outputDirectory,
-          '--require-epubcheck',
-        ],
-        {
-          encoding: 'utf8',
-          env: {
-            ...process.env,
-            EPUBCHECK_ARGUMENTS_LOG: argumentsLog,
-            PATH: `${binaryDirectory}:${process.env.PATH ?? ''}`,
-            SRT_PRIVATE_TEST_PDF: inputPath,
-          },
-          timeout: 120_000,
-        },
-      )
-      const receipt = JSON.parse(result.stdout)
-
-      expect(result.status, result.stderr).toBe(1)
-      expect(receipt).toMatchObject({
-        schemaVersion: '1.9.0',
-        execution: {
-          epubCheckRequired: true,
-          epubCheckPassedCount: 2,
-          allEpubCheckPassed: true,
-        },
-      })
-      expect(
-        receipt.runs
-          .flatMap((run) => run.artifacts)
-          .map((artifact) => artifact.epubCheck),
-      ).toEqual([{ status: 'passed' }, { status: 'passed' }])
-      expect(readdirSync(outputDirectory).sort()).toEqual([
-        'private-fidelity-receipt.json',
-        'run-1-mobile.epub',
-        'run-2-mobile.epub',
-      ])
-      expect(
-        readFileSync(argumentsLog, 'utf8').trim().split('\n'),
-      ).toHaveLength(2)
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  }, 120_000)
-
-  it('replays a hash-pinned empty decision set without exposing either local path', () => {
-    const temporaryDirectory = mkdtempSync(
-      join(tmpdir(), 'pdf-private-decisions-test-'),
-    )
-    const inputPath = resolve('tests/fixtures/pdf/born-digital.pdf')
-    const inputBytes = readFileSync(inputPath)
-    const inputSha256 = createHash('sha256').update(inputBytes).digest('hex')
-    const decisionPath = join(temporaryDirectory, 'owner decisions.json')
-    const decisionBytes = Buffer.from(
-      `${JSON.stringify({
-        schemaVersion: '1.1.0',
-        documentSha256: inputSha256,
-        decisions: [],
-      })}\n`,
-    )
-    const decisionSha256 = createHash('sha256')
-      .update(decisionBytes)
-      .digest('hex')
-    const outputDirectory = join(temporaryDirectory, 'proof')
-    try {
-      writeFileSync(decisionPath, decisionBytes, { mode: 0o600 })
-      const result = spawnSync(
-        process.execPath,
-        [
-          fileURLToPath(new URL('./pdf-private-fidelity.mjs', import.meta.url)),
-          '--input-env',
-          'SRT_PRIVATE_TEST_PDF',
-          '--paper-id',
-          'public-fixture',
-          '--expected-size',
-          String(inputBytes.byteLength),
-          '--expected-sha256',
-          inputSha256,
-          '--profiles',
-          'mobile',
-          '--repeat',
-          '2',
-          '--out',
-          outputDirectory,
-          '--decisions-env',
-          'SRT_PRIVATE_TEST_DECISIONS',
-          '--expected-decisions-sha256',
-          decisionSha256,
-        ],
-        {
-          encoding: 'utf8',
-          env: {
-            ...process.env,
-            SRT_PRIVATE_TEST_PDF: inputPath,
-            SRT_PRIVATE_TEST_DECISIONS: decisionPath,
-          },
-          timeout: 120_000,
-        },
-      )
-      const output = `${result.stdout}\n${result.stderr}`
-
-      expect(result.status, result.stderr).toBe(1)
-      expect(JSON.parse(result.stdout)).toMatchObject({
-        decisionSetSha256: decisionSha256,
-        execution: {
-          reconstructionDeterministic: true,
-          artifactsDeterministic: true,
-          localValidationPassed: true,
-        },
-        baselineComparison: { status: 'not-configured', passed: false },
-      })
-      expect(output).not.toContain(inputPath)
-      expect(output).not.toContain(decisionPath)
-      expect(output).not.toContain('owner decisions.json')
-    } finally {
-      rmSync(temporaryDirectory, { recursive: true, force: true })
-    }
-  }, 120_000)
 })
